Fix promise rejection from loadCostume and loadSound.
This fixes an issue where the costume loader would swallow errors when storage could not load a costume, resulting in a costume added to the target without the required information, causing a crash later.

Instead, reject the loadCostume promise so the error can be handled upstream. Within the VM, this means that the costume does not get added to the target, or that the sprite does not get created.